+++ src/plugins/paging/index.ts
@@ -4,6 +4,8 @@ export * from './basic/basicPaging.component';
 export * from './basic/basicPaging.interface';
 export * from './loadMore/loadMorePaging.component';
 export * from './loadMore/loadMorePaging.interface';
+export * from './previousNext/previousNextPaging.component';
+export * from './previousNext/previousNextPaging.interface';
 export * from './no/noPaging.component';
 export * from './no/noPaging.interface';
 export * from './plugins/pagingInitializer';